Move factories to testing package
